Fix for TKMHouse.ResTakeFromOut and delivery mode take-out
We should not allow to take redIn in this method, unless house is in take-out mode. 
We could try to take smth out via script action HouseTakeWaresFrom and thus try to get ware from In queue even if not in take-out mode. This will crash the game because we will try to remove offer as well