Actions.AIAttackAdd - its better to set delay from the gave start
We want to set and get attack parameters with the same state, without correlation to delay. 
So if player set attack with delay 10, then if he gets t back he also has to get delay 10, not 10 + gGame.Tick value.
MapMaker can calculate delay himself if he wants to add it to the current time.