Actions.OverlayTextAppend should also catch EConvertError
A.OverlayTextAppendFormatted method - this one would catch EConvertError raised by Delphi Format function, but by A.OverlayTextAppend - this one would not catch it, since on r13651 we just append plain text to the stored already Formatted string. But in r 14423 we append to the markUp string, which is going to be formatted with previously passed params by earlier calls of A.OverlayTextAppendFormatted.